renamed entity dir to model, added create controller for review, added contructors

+++ src/main/java/com/comp586/bonfilms/dao/FilmDetailRepository.java
@@ -1,6 +1,6 @@
 package com.comp586.bonfilms.dao;
 
-import com.comp586.bonfilms.entity.FilmDetail;
+import com.comp586.bonfilms.models.FilmDetail;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin("http://localhost:4200")
 @RepositoryRestResource(collectionResourceRel = "filmDetail", path = "film-detail")
-public interface FilmDetailRepository extends JpaRepository<FilmDetail, Long> {
+public interface FilmDetailRepository extends JpaRepository<FilmDetail, Integer> {
 }
